+++ lib/ui/page/home/widget/sharable.dart
@@ -17,43 +17,48 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:get/get.dart';
+import 'package:share_plus/share_plus.dart';
 
 import '/l10n/l10n.dart';
 import '/ui/widget/context_menu/menu.dart';
 import '/ui/widget/context_menu/region.dart';
-import '/ui/widget/svg/svg.dart';
 import '/ui/widget/text_field.dart';
 import '/util/message_popup.dart';
 
-/// Copyable text field that puts a [copy] of data into the clipboard on click
-/// or on context menu action.
-class CopyableTextField extends StatelessWidget {
-  const CopyableTextField({
+/// Sharable field opening a [Share] modal with the provided [share] content.
+class SharableTextField extends StatelessWidget {
+  SharableTextField({
     Key? key,
-    required this.state,
-    this.copy,
+    required String? text,
+    this.share,
     this.icon,
     this.label,
     this.style,
+    this.trailing,
     this.leading,
-  }) : super(key: key);
+  }) : super(key: key) {
+    state = TextFieldState(text: text, editable: false);
+  }
 
-  /// Reactive state of this [CopyableTextField].
-  final TextFieldState state;
+  /// Reactive state of this [SharableTextField].
+  late final TextFieldState state;
 
-  /// Data to put into the clipboard.
-  final String? copy;
+  /// Data to share or put into the clipboard.
+  final String? share;
 
   /// Optional leading icon.
   final IconData? icon;
 
+  /// Optional trailing [Widget].
+  final Widget? trailing;
+
   /// Optional leading [Widget].
   final Widget? leading;
 
-  /// Optional label of this [CopyableTextField].
+  /// Optional label of this [SharableTextField].
   final String? label;
 
-  /// [TextStyle] of this [CopyableTextField].
+  /// [TextStyle] of this [SharableTextField].
   final TextStyle? style;
 
   @override
@@ -71,7 +76,7 @@ class CopyableTextField extends StatelessWidget {
           ),
         Expanded(
           child: ContextMenuRegion(
-            enabled: (copy ?? state.text).isNotEmpty,
+            enabled: (share ?? state.text).isNotEmpty,
             decoration: BoxDecoration(borderRadius: BorderRadius.circular(25)),
             menu: ContextMenu(
               actions: [
@@ -83,21 +88,14 @@ class CopyableTextField extends StatelessWidget {
             ),
             child: InkWell(
               borderRadius: BorderRadius.circular(25),
-              onTap: (copy ?? state.text).isEmpty ? null : () => _copy(context),
+              onTap:
+                  (share ?? state.text).isEmpty ? null : () => _share(context),
               child: IgnorePointer(
                 child: ReactiveTextField(
                   prefix: leading,
                   state: state,
-                  trailing: Transform.translate(
-                    offset: const Offset(0, -1),
-                    child: Transform.scale(
-                      scale: 1.15,
-                      child: SvgLoader.asset(
-                        'assets/icons/copy.svg',
-                        height: 15,
-                      ),
-                    ),
-                  ),
+                  suffix: trailing == null ? Icons.ios_share : null,
+                  trailing: trailing,
                   label: label,
                   style: style,
                 ),
@@ -109,9 +107,25 @@ class CopyableTextField extends StatelessWidget {
     );
   }
 
-  /// Puts a [copy] of data into the clipboard and shows a snackbar.
+  /// Puts the [share] into the clipboard and shows a snackbar.
   void _copy(BuildContext context) {
-    Clipboard.setData(ClipboardData(text: copy ?? state.text));
+    Clipboard.setData(ClipboardData(text: share ?? state.text));
     MessagePopup.success('label_copied_to_clipboard'.l10n);
+  }
+
+  /// Opens a [Share] modal sharing the [share].
+  Future<void> _share(BuildContext context) async {
+    Rect? rect;
+
+    try {
+      final box = context.findRenderObject() as RenderBox?;
+      if (box != null) {
+        rect = box.localToGlobal(Offset.zero) & box.size;
+      }
+    } catch (e) {
+      // No-op.
+    }
+
+    await Share.share(share ?? state.text, sharePositionOrigin: rect);
   }
 }